requirements.txt
appdirs==1.4.3
Appium-Python-Client==0.26
beautifulsoup4==4.6.0
etiqa-behave==1.2.6.0
enum34==1.1.6
future==0.16.0
linecache2==1.0.0
lxml==3.8.0
olefile==0.44
packaging==16.8
parse==1.8.0
parse-type==0.3.4
Pillow==4.1.1
PyHamcrest==1.8.5
pyparsing==2.2.0
etiqa-pytractor==0.2.1.1
selenium==3.6.0
six==1.10.0
traceback2==1.4.0
unittest-xml-reporting==2.1.0
wheel
